vlm-caption-finetune-mlops/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-caption-finetune-mlops"
version = "0.1.0"
description = "Fine-tune Vision-Language Models for image captioning with MLOps"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["vlm", "vision-language", "captioning", "mlops", "fine-tuning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "accelerate>=0.28.0",
    "datasets>=2.18.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.10.0",
    "wandb>=0.16.0",
    "dvc>=3.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gradio>=4.0.0",
    "nltk>=3.8.0",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.4.0",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
inference = [
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
]

[project.urls]
Repository = "https://github.com/your-org/vlm-caption-finetune-mlops"
Documentation = "https://github.com/your-org/vlm-caption-finetune-mlops#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
